TXDK/sunobrainapi/pyproject.toml
[project]
name = "sunobrainapi"
version = "1.0.0"
description = "SunoBrain API — Suno v5.5 lyric optimization engine powered by Gemini"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.0.0",
    "google-genai>=1.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
